Added profile creation and fixed user delete

+++ src/users/users.service.ts
@@ -1,14 +1,17 @@
-import { BadRequestException, Injectable } from "@nestjs/common";
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { User } from "./users.entity";
+import { User } from "./entities/users.entity";
 import { Repository } from "typeorm";
-import { UserDTO } from "./users.dto";
+import { ProfileDTO, UserDTO } from "./users.dto";
 import { HasherService } from "src/hasher/hasher.service";
+import { Profile } from "./entities/profiles.entity";
+import { DeleteResult } from "typeorm/browser";
 
 @Injectable()
 export class UsersService {
     constructor(
         @InjectRepository(User) private userRepository: Repository<User>,
+        @InjectRepository(Profile) private profileRepository: Repository<Profile>,
         private hasherService: HasherService
     ) { }
 
@@ -32,7 +35,46 @@ export class UsersService {
         return this.userRepository.save(newUser);
     }
 
-    delete(user_id: number) {
-        return this.userRepository.delete({ id: user_id })
+    async createProfile(user_id: number, profile: ProfileDTO) {
+        const user = await this.userRepository.findOne({
+            where: { id: user_id },
+            relations: ['profile']
+        });
+        if (!user) {
+            throw new NotFoundException('No such user exists');
+        }
+        if (user.profile) {
+            throw new BadRequestException('User already has a profile');
+        }
+
+        const newProfile = this.profileRepository.create(profile);
+        user.profile = newProfile;
+
+        return this.userRepository.save(user);
+    }
+
+    async delete(user_id: number) {
+        const user = await this.userRepository.findOne({
+            where: { id: user_id },
+            relations: {
+                profile: true
+            }
+        })
+
+        if (!user) {
+            return { affected: 0 }
+        }
+
+        let result: Promise<DeleteResult>;
+        if (user.profile) {
+            const profile_id = user.profile.id;
+            user.profile = null;
+            result = this.userRepository.delete({ id: user_id });
+            await this.profileRepository.delete({ id: profile_id });
+        } else {
+            result = this.userRepository.delete({ id: user_id });
+        }
+
+        return result;
     }
 }
